+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/MythicMobManager.java
@@ -1,7 +1,9 @@
 package com.elmakers.mine.bukkit.integration;
 
+import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.Optional;
+import java.util.UUID;
 import javax.annotation.Nullable;
 
 import org.bukkit.Location;
@@ -10,15 +12,18 @@ import org.bukkit.plugin.Plugin;
 
 import com.elmakers.mine.bukkit.magic.MagicController;
 
-import io.lumine.xikage.mythicmobs.MythicMobs;
-import io.lumine.xikage.mythicmobs.mobs.ActiveMob;
+import io.lumine.mythic.api.MythicPlugin;
+import io.lumine.mythic.api.mobs.MobManager;
+import io.lumine.mythic.api.mobs.MythicMob;
+import io.lumine.mythic.bukkit.BukkitAdapter;
+import io.lumine.mythic.core.mobs.ActiveMob;
 
 public class MythicMobManager {
 
     private final MagicController controller;
     private final Plugin plugin;
 
-    private MythicMobs api = null;
+    private MythicPlugin api = null;
 
     public MythicMobManager(MagicController controller, Plugin plugin) {
         this.controller = controller;
@@ -26,12 +31,12 @@ public class MythicMobManager {
     }
 
     public boolean initialize() {
-        if (plugin == null || !(plugin instanceof MythicMobs)) {
+        if (plugin == null || !(plugin instanceof MythicPlugin)) {
             return false;
         }
 
-        api = MythicMobs.inst();
-        return api != null;
+        api = (MythicPlugin)plugin;
+        return true;
     }
 
     public boolean isEnabled() {
@@ -39,8 +44,14 @@ public class MythicMobManager {
     }
 
     @Nullable
-    public Entity spawn(String key, Location location) {
-        ActiveMob mob = api.getMobManager().spawnMob(key, location);
+    public Entity spawn(String key, Location location, double level) {
+        if (api == null) return null;
+        Optional<MythicMob> mythicMob = api.getMobManager().getMythicMob(key);
+        if (!mythicMob.isPresent()) {
+            controller.getLogger().warning("Unknown mythic mob type: " + key);
+            return null;
+        }
+        ActiveMob mob = mythicMob.get().spawn(BukkitAdapter.adapt(location), level);
         if (mob == null) {
             controller.getLogger().warning("Unable to spawn mythic mob with id of " + key);
             return null;
@@ -50,19 +61,21 @@ public class MythicMobManager {
     }
 
     public Collection<String> getMobKeys() {
+        if (api == null) return null;
         return api.getMobManager().getMobNames();
     }
 
     public boolean isMobKey(String mobKey) {
+        if (api == null) return false;
         // Hopefully this is backed by a Set?
         return api.getMobManager().getMobNames().contains(mobKey);
     }
 
     public void setMobLevel(Entity entity, double level) {
-        if (entity == null) {
+        if (api == null || entity == null) {
             return;
         }
-        Optional<ActiveMob> mob = api.getMobManager().getActiveMob(entity.getUniqueId());
+        Optional<ActiveMob> mob = getActiveMob(entity.getUniqueId());
         if (!mob.isPresent()) {
             return;
         }
@@ -71,10 +84,10 @@ public class MythicMobManager {
 
     @Nullable
     public Double getMobLevel(Entity entity) {
-        if (entity == null) {
+        if (api == null || entity == null) {
             return null;
         }
-        Optional<ActiveMob> mob = api.getMobManager().getActiveMob(entity.getUniqueId());
+        Optional<ActiveMob> mob = getActiveMob(entity.getUniqueId());
         if (!mob.isPresent()) {
             return null;
         }
@@ -83,13 +96,32 @@ public class MythicMobManager {
 
     @Nullable
     public String getMobKey(Entity entity) {
-        if (entity == null) {
+        if (api == null || entity == null) {
             return null;
         }
-        Optional<ActiveMob> mob = api.getMobManager().getActiveMob(entity.getUniqueId());
+        Optional<ActiveMob> mob = getActiveMob(entity.getUniqueId());
         if (!mob.isPresent()) {
             return null;
         }
         return mob.get().getMobType();
+    }
+
+    // Not in the API...
+    @SuppressWarnings("unchecked")
+    public Optional<ActiveMob> getActiveMob(UUID id) {
+        try {
+            MobManager manager = api.getMobManager();
+            Method getActiveMobMethod = manager.getClass().getMethod("getActiveMob", UUID.class);
+            if (getActiveMobMethod != null) {
+                return (Optional<ActiveMob>)getActiveMobMethod.invoke(manager, id);
+            } else {
+                controller.getLogger().warning("MythicMobs integration has gone wrong, disabling");
+                api = null;
+            }
+        } catch (Exception ex) {
+            controller.getLogger().warning("MythicMobs integration has gone wrong, disabling");
+            api = null;
+        }
+        return Optional.empty();
     }
 }