+++ Magic/CHANGELOG.md
@@ -16,6 +16,9 @@
 
 ## Fixes
 
+ - Update to latest MythicMobs API
+   Note that because the Mythic API changed completely, you will need to update MythicMobs when you update Magic.  
+   Untested.
  - Fix boss bar display for spells like Cloak when switching to a different class while the spell is active
  - Fix issue updating mobs that have despawned on config reload